api: create a v2 endpoint for recovering backups
