+++ api/views/mixins.py
@@ -1,5 +1,6 @@
 from inspect import cleandoc
 from drf_spectacular.utils import extend_schema
+from mimetypes import guess_type
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -7,8 +8,8 @@ from lib import backup_helper
 from lib.auth import authorized
 
 from anthias_app.models import Asset
-from os import remove
-from settings import settings
+from os import path, remove
+from settings import settings, ZmqPublisher
 
 
 class DeleteAssetViewMixin:
@@ -51,3 +52,50 @@ class BackupViewMixin(APIView):
     def post(self, request):
         filename = backup_helper.create_backup(name=settings['player_name'])
         return Response(filename, status=status.HTTP_201_CREATED)
+
+
+class RecoverViewMixin(APIView):
+    @extend_schema(
+        summary='Recover from backup',
+        description=cleandoc("""
+        Recover data from a backup file. The backup file must be
+        a `.tar.gz` file.
+        """),
+        request={
+            'multipart/form-data': {
+                'type': 'object',
+                'properties': {
+                    'backup_upload': {
+                        'type': 'string',
+                        'format': 'binary'
+                    }
+                }
+            }
+        },
+        responses={
+            200: {
+                'type': 'string',
+                'example': 'Recovery successful.',
+            }
+        },
+    )
+    @authorized
+    def post(self, request):
+        publisher = ZmqPublisher.get_instance()
+        file_upload = (request.data.get('backup_upload'))
+        filename = file_upload.name
+
+        if guess_type(filename)[0] != 'application/x-tar':
+            raise Exception("Incorrect file extension.")
+        try:
+            publisher.send_to_viewer('stop')
+            location = path.join("static", filename)
+
+            with open(location, 'wb') as f:
+                f.write(file_upload.read())
+
+            backup_helper.recover(location)
+
+            return Response("Recovery successful.")
+        finally:
+            publisher.send_to_viewer('play')
